requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
websockets==13.0

twilio==9.2.3
python-dotenv==1.1.0

httpx==0.28.1
openai==1.55.3
